requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-decouple==3.8
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
pydantic==2.12.5
pydantic[email]==2.12.5
openai==1.57.0
python-multipart==0.0.20
gunicorn==21.2.0

